wb-cardgen/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
pydantic-settings==2.4.0
python-multipart==0.0.9
Pillow==10.4.0
numpy==1.26.4
rembg==2.0.56
starlette==0.37.2
